backend/requirements.txt
# Base
Django==3.0.5
uvicorn==0.11.5
gunicorn==20.0.4
psycopg2-binary==2.8.5
whitenoise==5.0.1

# Debug
Werkzeug==1.0.1
django-extensions==2.2.9
ptipython==1.0.1
ptvsd==4.3.2
django-debug-toolbar==2.2

# Worker
celery==4.4.2
django-redis==4.11.0
redis==3.5.0

# REST
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dj-rest-auth==1.0.4
rest-condition==1.0.3
drf-nested-routers==0.91

# Monitoring
django-prometheus==2.0.0

# Math
numpy==1.18.3
scipy==1.4.1

# Lint and type
black==19.10b0
pylint==2.5.0
pylint-django==2.0.15
